ML_PROJECT/ML_PROJECT/requirements.txt
langgraph==0.2.55
langchain==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.3
langchain-openai==0.2.14
langchain-anthropic==0.3.3
python-dotenv==1.0.1
scikit-learn==1.5.2
typing-extensions==4.12.2
